+++ packages/react/cypress/support/commands.ts
@@ -2,7 +2,6 @@
 /// <reference types="cypress" />
 
 import { CYPRESS_AXE_OPTIONS, CYPRESS_AXE_REPORTING } from "../utils/a11y";
-import "cypress-audit/commands";
 
 import {
   BE_VISIBLE,